Allow public access to proxy_t c_ptr() and state
There are valid use cases for clients wanting to use the
underlying wl_proxy* of a proxy_t and there should be no particular
reason to deny them access to it. Also, since the proxy_t knows whether
it currently wraps any object, this should also be made available to
clients. Functionality for explicitly releasing the proxy can also
be useful.
The new functions have the prefix "proxy_" in order to minimize the
chances of name clashes with Wayland interface functions.