Use versioned constructor function for registry_t::bind()
Since Wayland protocol versions on the client are inherited from their
parent objects, the version of all objects will be 0 if the root
objects received from the compositor with registry_t::bind() do not
have a version set. For this to happen,
wl_proxy_marshal_array_constructor_versioned() must be used.