Fix handling of objects that are owned by other libraries
Some objects (most notably: buffers from libwayland-cursor) are created
in other libraries and we just wrap them so they can be used as
arguments.

Previously, they would have the special parameter dontdestroy set so
that they would not be destroyed if the waylandpp proxy goes out of
scope since the library that owns them would get into trouble then.

But this is not enough: If the waylandpp proxy is destroyed and events
happen after that, the original library will not get them since we
have added our dispatcher. And worse than that, our dispatcher will
crash since the C++ wrapper does not exist any more.

So to fix this, change "dontdestroy" to proper "foreign": If foreign is
set, it means that another library is responsible and we should neither
add our dispatcher nor destroy the object nor set any user data.