link to current libraries
Some of the libraries need to link to the just built libraries of the
project, instead of stale versions that may exist in the system.

Since the helper function define_library does not support additional
arguments due to the usage of ARGN, append the required
target_link_libraries call to make cmake aware of the build dependency.

Signed-off-by: Olaf Hering <olaf@aepfle.de>